[epub_view] improved generator and refactored structure
