+++ packages/epub_view/lib/src/epub_view.dart
@@ -1,8 +1,8 @@
 import 'dart:async';
 
 import 'package:epub/epub.dart';
-import 'package:epub_view/src/parser/epub_cfi.dart';
-import 'package:epub_view/src/generator/epub_cfi.dart';
+import 'package:epub_view/src/epub_cfi/parser.dart';
+import 'package:epub_view/src/epub_cfi/generator.dart';
 import 'package:html/parser.dart' show parse;
 import 'package:flutter/widgets.dart';
 import 'package:flutter_html/flutter_html.dart';
